requirements.txt
albumentations==1.3.1
diffsort==0.2.0
fast_slic==0.4.0
matplotlib==3.7.1
numba==0.57.1
numpy==1.24.3
opencv_python==4.7.0.72
opencv_python_headless==4.7.0.72
palettable==3.3.3
pandas==2.0.1
Pillow==10.3.0
scikit_learn==1.2.2
scipy==1.13.0
scikit-image
timm==0.9.7
torch==2.1.0
torch_optimizer==0.3.0
torchsummary==1.5.1
torchvision==0.16.0
tqdm==4.65.0
datasets
